requirements.txt
Django==4.2.9
djangorestframework==3.14.0
django-cors-headers==4.3.1
Pillow==10.1.0
markdown==3.5.1
django-filter==23.5
python-dateutil==2.8.2
